@@ -3,7 +3,9 @@ package christopher.popularmovies.Data;
 
 import android.os.Handler;
 import android.os.Looper;
-import android.support.annotation.NonNull;
+
+
+import androidx.annotation.NonNull;
 
 import java.util.concurrent.Executor;
 import java.util.concurrent.Executors;
