@@ -1,6 +1,8 @@
 package christopher.popularmovies.Data;
 
-import android.arch.persistence.room.TypeConverter;
+
+
+import androidx.room.TypeConverter;
 
 import java.util.Date;
 
